feat: new project modal
