--- src/presentation/components/InfoProjectModal/style.ts
@@ -1,19 +1,0 @@
-import styled from "styled-components";
-
-export const ModalInfoContent = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 1rem;
-
-  text-align: center;
-
-  strong {
-    font-size: 1.2rem;
-  }
-
-  p {
-    color: gray;
-  }
-`;
