+++ src/presentation/components/Modal/InfoProjectModal/index.tsx
@@ -1,8 +1,8 @@
-import { Project } from "../../../data/interfaces/project";
-import { withCopy } from "../../hooks/withCopy";
-import { Avatar } from "../Avatar";
-import { MembersList } from "../MembersList";
-import { Tag } from "../Tag";
+import { Project } from "../../../../data/interfaces/project";
+import { withCopy } from "../../../hooks/withCopy";
+import { Avatar } from "../../Avatar";
+import { MembersList } from "../../MembersList";
+import { Tag } from "../../Tag";
 import { ModalInfoContent } from "./style";
 
 type Props = {
@@ -12,7 +12,7 @@ type Props = {
 export const InfoProjectModal = ({project}: Props) => {
 
   return (
-    <ModalInfoContent>
+    <ModalInfoContent gap="8px">
       <Avatar src={project.owner.photoUrl} alt={project.owner.name + " photo"}/>
       <strong>{project.title}</strong>
       {withCopy(<Tag size='small' label={project.id ?? ""} />, project.id ?? "")}
