+++ src/presentation/components/Modal/InfoProjectModal/style.ts
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+import { Column } from "../../Layouts/Column";
+
+export const ModalInfoContent = styled(Column)`
+  text-align: center;
+
+  strong {
+    font-size: 1.2rem;
+  }
+
+  p {
+    color: gray;
+  }
+`;
