+++ src/presentation/components/Modal/NewProjectModal/style.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+import { Column } from "../../Layouts/Column";
+
+
+export const ModalTaskContent = styled(Column)`
+  #text-button {
+    color: var(--white);
+    font-size: 14px;
+    font-weight: bold;
+  }
+`;
